requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
scipy>=1.10.0
pywavelets>=1.4.0
pyyaml>=6.0.0
pytest>=7.0.0